% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{defaults-package}
\alias{defaults-package}
\title{defaults}
\description{
defaults makes it easy to change the default values for function
  arguments, whether they are in packages, or defined locally. You just need
  the function \code{\link{defaults}}.
}
