% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{defaults<-}
\alias{reset_defaults}
\title{change a function's default arguments}
\usage{
defaults(fun)

defaults(fun) <- value

reset_defaults(fun)
}
\arguments{
\item{fun}{a function}

\item{value}{a named list of new default arguments for that function}
}
\description{
\code{defaults()} lets you check, and change, a function's
  default arguments, \code{reset_defaults()} returns the arguments to their
  original defaults.
}
\details{
If \code{fun} is in a package, a function of the same name will be
  defined in the calling environment (probably your workspace). If \code{fun}
  is defined locally, it will be overwritten by the version with the new
  defaults.

  \code{reset_defaults} returns the reset function, rather than modifying it
  in place, so you'll need to reassign it as in the example.
}
\examples{
# to do
}
